Optimize code by lazy loading components, removing unused fonts, and enhancing configuration for better performance
