+++ config/fonts.ts
@@ -2,7 +2,6 @@ import {
   Fira_Code as FontMono,
   Inter as FontSans,
   Playfair_Display,
-  Song_Myung,
 } from "next/font/google";
 
 export const fontSans = FontSans({
@@ -22,11 +21,5 @@ export const fontPlayfair = Playfair_Display({
   variable: "--font-playfair",
   weight: ["400", "500", "600", "700"],
   style: ["normal", "italic"],
-  display: "swap",
-});
-
-export const fontSongMyung = Song_Myung({
-  variable: "--font-songmyung",
-  weight: "400",
   display: "swap",
 });
