+++ package.json
@@ -9,6 +9,7 @@
     "dev:https": "next dev --experimental-https",
     "prebuild": "node convert-images.mjs ./public && node optimize-svgs.mjs ./public",
     "build": "next build",
+    "analyze": "ANALYZE=true pnpm build",
     "start": "next start",
     "lint": "eslint --fix",
     "typecheck": "tsc --noEmit",
